[REFACTOR] DTO 클래스 정리
