+++ src/main/java/tf/tailfriend/reserve/controller/PaymentController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import tf.tailfriend.global.service.DateTimeRange;
+import tf.tailfriend.reserve.dto.DateTimeRange;
 import tf.tailfriend.reserve.dto.ListResponseDto;
 import tf.tailfriend.reserve.dto.PaymentInfoResponseDto;
 import tf.tailfriend.reserve.dto.PaymentListRequestDto;
