+++ src/main/java/tf/tailfriend/reserve/dto/PaymentListRequestDto.java
@@ -1,11 +1,6 @@
 package tf.tailfriend.reserve.dto;
 
 import lombok.*;
-import org.springframework.format.annotation.DateTimeFormat;
-import tf.tailfriend.global.service.DateTimeRange;
-import tf.tailfriend.reserve.entity.Payment;
-
-import java.time.LocalDateTime;
 
 @Getter
 @Builder
