+++ src/main/java/tf/tailfriend/reserve/repository/PaymentDao.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import tf.tailfriend.reserve.entity.Payment;
 
-import java.util.List;
-
 @Repository
 public interface PaymentDao extends JpaRepository<Payment, Integer>, CustomPaymentDao {
 }
