+++ src/main/resources/application.yml
@@ -19,7 +19,7 @@ spring:
     driver-class-name: com.mysql.cj.jdbc.Driver
     username: ${DB_USERNAME}
     password: ${DB_PASSWORD}
-    
+
     hikari:
       maximum-pool-size: 10
       minimum-idle: 2
@@ -47,7 +47,7 @@ spring:
               - account_email
             client-authentication-method: client_secret_post
             client-name: Kakao
-          
+
           naver:
             authorization-grant-type: authorization_code
             client-id: ${NAVER_CLIENT_ID}
@@ -57,7 +57,7 @@ spring:
               - account_email
             client-authentication-method: client_secret_post
             client-name: Naver
-          
+
           google:
             authorization-grant-type: authorization_code
             client-id: ${GOOGLE_CLIENT_ID}
@@ -67,26 +67,26 @@ spring:
               - email
             client-authentication-method: client_secret_post
             client-name: Google
-        
+
         provider:
           kakao:
             authorization-uri: https://kauth.kakao.com/oauth/authorize
             token-uri: https://kauth.kakao.com/oauth/token
             user-info-uri: https://kapi.kakao.com/v2/user/me
             user-name-attribute: id
-          
+
           naver:
             authorization-uri: https://nid.naver.com/oauth2.0/authorize
             token-uri: https://nid.naver.com/oauth2.0/token
             user-info-uri: https://openapi.naver.com/v1/nid/me
             user-name-attribute: response
-          
+
           google:
             authorization-uri: https://accounts.google.com/o/oauth2/auth
             token-uri: https://oauth2.googleapis.com/token
             user-info-uri: https://www.googleapis.com/oauth2/v3/userinfo
             user-name-attribute: sub
-  
+
 jwt:
   secret: ${JWT_SECRET}
   token-validity: 86400
@@ -107,4 +107,3 @@ ncp:
   access-key: ${NCR_USERNAME}
   secret-key: ${NCR_PASSWORD}
   bucket-name: ${NCP_BUCKET_NAME}
-
